+++ packages/datadog_flutter_plugin/e2e_test_app/integration_test/rum/rum_monitor_test.dart
@@ -325,7 +325,7 @@ void main() {
 
     await measure('flutter_stop_resource_with_error', () {
       datadog.rum!.stopResourceLoadingWithErrorInfo(
-          resourceKey, randomString(), e2eAttributes(tester));
+          resourceKey, randomString(), randomString(), e2eAttributes(tester));
     });
 
     datadog.rum!.stopView(viewKey);