+++ packages/datadog_flutter_plugin/lib/src/datadog_configuration.dart
@@ -2,6 +2,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-2022 Datadog, Inc.
 
+import '../datadog_flutter_plugin.dart';
+
 /// Defines the Datadog SDK policy when batching data together before uploading
 /// it to Datadog servers. Smaller batches mean smaller but more network
 /// requests, whereas larger batches mean fewer but larger network requests.
@@ -89,6 +91,11 @@ class LoggingConfiguration {
   /// Defaults to `false`.
   bool printLogsToConsole;
 
+  // Enables or disables sending logs to Datadog.
+  ///
+  /// Defaults to `true`.
+  bool sendLogsToDatadog;
+
   /// Enables the logs integration with RUM.
   ///
   /// If enabled, all the logs will be enriched with the current RUM View
@@ -106,19 +113,29 @@ class LoggingConfiguration {
   /// Default to `true`.
   bool bundleWithTrace;
 
+  /// Sets the name of the logger.
+  ///
+  /// This name will be set as the `logger.name` attribute attached to all logs
+  /// sent to Datadog from this logger.
+  String? loggerName;
+
   LoggingConfiguration({
     this.sendNetworkInfo = false,
     this.printLogsToConsole = false,
+    this.sendLogsToDatadog = true,
     this.bundleWithRum = true,
     this.bundleWithTrace = true,
+    this.loggerName,
   });
 
   Map<String, dynamic> encode() {
     return {
       'sendNetworkInfo': sendNetworkInfo,
       'printLogsToConsole': printLogsToConsole,
+      'sendLogsToDatadog': sendLogsToDatadog,
       'bundleWithRum': bundleWithRum,
       'bundleWithTrace': bundleWithTrace,
+      'loggerName': loggerName,
     };
   }
 }
@@ -192,8 +209,9 @@ class DdSdkConfiguration {
   /// Whether or not to enable native crash reporting.
   bool nativeCrashReportEnabled;
 
-  /// The [DatadogSite] to send information to.
-  DatadogSite? site;
+  /// The [DatadogSite] to send information to. This site must match the site
+  /// used to generate your client token.
+  DatadogSite site;
 
   /// The initial [TrackingConsent] for this user.
   TrackingConsent trackingConsent;
@@ -208,28 +226,6 @@ class DdSdkConfiguration {
 
   /// Set a custom endpoint to send information to.
   String? customEndpoint;
-
-  /// Configures network requests monitoring for Tracing and RUM features.
-  ///
-  /// If set, the SDK will override [HttpClient] creation (via [HttpOverrides])
-  /// to provide its own implementation. For more information, check the
-  /// documentation on [DatadogTrackingHttpClient]
-  ///
-  /// If the RUM feature is enabled, the SDK will send RUM Resources for all
-  /// intercepted requests.
-  ///
-  /// If the Tracing feature is enabled, the SDK will send tracing Span for each
-  /// 1st-party request. It will also add extra HTTP headers to further
-  /// propagate the trace - it means that if your backend is instrumented with
-  /// Datadog agent you will see the full trace (e.g.: client → server →
-  /// database) in your dashboard, thanks to Datadog Distributed Tracing.
-  ///
-  /// If both RUM and Tracing features are enabled, the SDK will be sending RUM
-  /// Resources for 1st- and 3rd-party requests and tracing Spans for
-  /// 1st-parties.
-  ///
-  /// See also [firstPartyHosts]
-  bool trackHttpClient;
 
   /// A list of first party hosts, used in conjunction with [trackHttpClient]
   ///
@@ -263,33 +259,38 @@ class DdSdkConfiguration {
   /// Any additional configuration to be passed to the native SDKs
   final Map<String, dynamic> additionalConfig = {};
 
+  /// Configurations for additional plugins that will be created after Datadog
+  /// is initialized.
+  final List<DatadogPluginConfiguration> additionalPlugins = [];
+
   DdSdkConfiguration({
     required this.clientToken,
     required this.env,
     required this.trackingConsent,
+    required this.site,
     this.nativeCrashReportEnabled = false,
-    this.site,
     this.uploadFrequency,
     this.batchSize,
     this.customEndpoint,
-    this.trackHttpClient = false,
     this.firstPartyHosts = const [],
     this.loggingConfiguration,
     this.tracingConfiguration,
     this.rumConfiguration,
   });
 
+  void addPlugin(DatadogPluginConfiguration pluginConfiguration) =>
+      additionalPlugins.add(pluginConfiguration);
+
   Map<String, dynamic> encode() {
     return {
       'clientToken': clientToken,
       'env': env,
       'nativeCrashReportEnabled': nativeCrashReportEnabled,
-      'site': site?.toString(),
+      'site': site.toString(),
       'batchSize': batchSize?.toString(),
       'uploadFrequency': uploadFrequency?.toString(),
       'trackingConsent': trackingConsent.toString(),
       'customEndpoint': customEndpoint,
-      'loggingConfiguration': loggingConfiguration?.encode(),
       'tracingConfiguration': tracingConfiguration?.encode(),
       'rumConfiguration': rumConfiguration?.encode(),
       'additionalConfig': additionalConfig