+++ packages/datadog_flutter_plugin/test/logs/ddlogs_method_channel_test.dart
@@ -2,6 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-Present Datadog, Inc.
 
+import 'package:datadog_flutter_plugin/datadog_flutter_plugin.dart';
 import 'package:datadog_flutter_plugin/src/logs/ddlogs_method_channel.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -24,11 +25,24 @@ void main() {
     log.clear();
   });
 
+  test('create logger passed to method channel', () async {
+    final config = LoggingConfiguration(loggerName: 'loggerName');
+    await ddLogsPlatform.createLogger('uuid', config);
+
+    expect(log, <Matcher>[
+      isMethodCall('createLogger', arguments: {
+        'loggerHandle': 'uuid',
+        'configuration': config.encode(),
+      })
+    ]);
+  });
+
   test('debug logs passed to method channel', () async {
-    await ddLogsPlatform.debug('debug message', {'attribute': 'value'});
+    await ddLogsPlatform.debug('uuid', 'debug message', {'attribute': 'value'});
 
     expect(log, <Matcher>[
       isMethodCall('debug', arguments: {
+        'loggerHandle': 'uuid',
         'message': 'debug message',
         'context': {'attribute': 'value'}
       })
@@ -36,10 +50,11 @@ void main() {
   });
 
   test('info logs passed to method channel', () async {
-    await ddLogsPlatform.info('info message', {'attribute': 'value'});
+    await ddLogsPlatform.info('uuid', 'info message', {'attribute': 'value'});
 
     expect(log, <Matcher>[
       isMethodCall('info', arguments: {
+        'loggerHandle': 'uuid',
         'message': 'info message',
         'context': {'attribute': 'value'}
       })
@@ -47,10 +62,11 @@ void main() {
   });
 
   test('warn logs passed to method channel', () async {
-    await ddLogsPlatform.warn('warn message', {'attribute': 'value'});
+    await ddLogsPlatform.warn('uuid', 'warn message', {'attribute': 'value'});
 
     expect(log, <Matcher>[
       isMethodCall('warn', arguments: {
+        'loggerHandle': 'uuid',
         'message': 'warn message',
         'context': {'attribute': 'value'}
       })
@@ -58,10 +74,11 @@ void main() {
   });
 
   test('error logs passed to method channel', () async {
-    await ddLogsPlatform.error('error message', {'attribute': 'value'});
+    await ddLogsPlatform.error('uuid', 'error message', {'attribute': 'value'});
 
     expect(log, <Matcher>[
       isMethodCall('error', arguments: {
+        'loggerHandle': 'uuid',
         'message': 'error message',
         'context': {'attribute': 'value'}
       })
@@ -69,17 +86,20 @@ void main() {
   });
 
   test('addAttribute passed to method channel', () async {
-    await ddLogsPlatform.addAttribute('my_key', 'my_value');
+    await ddLogsPlatform.addAttribute('uuid', 'my_key', 'my_value');
 
     expect(log, <Matcher>[
-      isMethodCall('addAttribute',
-          arguments: {'key': 'my_key', 'value': 'my_value'})
+      isMethodCall('addAttribute', arguments: {
+        'loggerHandle': 'uuid',
+        'key': 'my_key',
+        'value': 'my_value',
+      })
     ]);
   });
 
   test('addAttributes passes complicated values to method channel', () async {
-    await ddLogsPlatform.addAttribute('my_attribute', true);
-    await ddLogsPlatform.addAttribute('my_attribute', {
+    await ddLogsPlatform.addAttribute('uuid', 'my_attribute', true);
+    await ddLogsPlatform.addAttribute('uuid', 'my_attribute', {
       'int_value': 256,
       'bool_value': false,
       'double_value': 2.3,
@@ -87,9 +107,13 @@ void main() {
     });
 
     expect(log, <Matcher>[
-      isMethodCall('addAttribute',
-          arguments: {'key': 'my_attribute', 'value': true}),
       isMethodCall('addAttribute', arguments: {
+        'loggerHandle': 'uuid',
+        'key': 'my_attribute',
+        'value': true,
+      }),
+      isMethodCall('addAttribute', arguments: {
+        'loggerHandle': 'uuid',
         'key': 'my_attribute',
         'value': {
           'int_value': 256,
@@ -102,44 +126,59 @@ void main() {
   });
 
   test('removeAttribute passes to method channel', () async {
-    await ddLogsPlatform.removeAttribute('my_attribute');
+    await ddLogsPlatform.removeAttribute('uuid', 'my_attribute');
 
     expect(log, <Matcher>[
       isMethodCall('removeAttribute', arguments: {
+        'loggerHandle': 'uuid',
         'key': 'my_attribute',
       })
     ]);
   });
 
   test('addTag passes tag to method channel', () async {
-    await ddLogsPlatform.addTag('my_tag');
+    await ddLogsPlatform.addTag('uuid', 'my_tag');
 
     expect(log, <Matcher>[
-      isMethodCall('addTag', arguments: {'tag': 'my_tag', 'value': null})
+      isMethodCall('addTag', arguments: {
+        'loggerHandle': 'uuid',
+        'tag': 'my_tag',
+        'value': null,
+      })
     ]);
   });
 
   test('addTag passes tag and value to method channel', () async {
-    await ddLogsPlatform.addTag('my_tag', 'tag_value');
+    await ddLogsPlatform.addTag('uuid', 'my_tag', 'tag_value');
 
     expect(log, <Matcher>[
-      isMethodCall('addTag', arguments: {'tag': 'my_tag', 'value': 'tag_value'})
+      isMethodCall('addTag', arguments: {
+        'loggerHandle': 'uuid',
+        'tag': 'my_tag',
+        'value': 'tag_value',
+      })
     ]);
   });
 
   test('removeTag passes tag to method channel', () async {
-    await ddLogsPlatform.removeTag('my_tag');
+    await ddLogsPlatform.removeTag('uuid', 'my_tag');
 
     expect(log, <Matcher>[
-      isMethodCall('removeTag', arguments: {'tag': 'my_tag'}),
+      isMethodCall('removeTag', arguments: {
+        'loggerHandle': 'uuid',
+        'tag': 'my_tag',
+      }),
     ]);
   });
 
   test('removeTagWithKey passed to method channel', () async {
-    await ddLogsPlatform.removeTagWithKey('my_tag');
+    await ddLogsPlatform.removeTagWithKey('uuid', 'my_tag');
 
     expect(log, <Matcher>[
-      isMethodCall('removeTagWithKey', arguments: {'key': 'my_tag'})
+      isMethodCall('removeTagWithKey', arguments: {
+        'loggerHandle': 'uuid',
+        'key': 'my_tag',
+      })
     ]);
   });
 }