+++ packages/datadog_tracking_http_client/lib/src/tracking_http_client.dart
@@ -2,15 +2,16 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-2022 Datadog, Inc.
 
+// ignore_for_file: invalid_use_of_internal_member
+
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:datadog_flutter_plugin/datadog_flutter_plugin.dart';
 import 'package:uuid/uuid.dart';
 
-import '../datadog_flutter_plugin.dart';
-import 'attributes.dart';
-import 'rum/ddrum.dart';
+import 'tracing_headers.dart';
 
 /// Overrides to supply the [DatadogTrackingHttpClient] instead of the default
 /// HttpClient
@@ -117,6 +118,7 @@ class DatadogTrackingHttpClient implements HttpClient {
           attributes[DatadogPlatformAttributeKey.spanID] = spanId;
 
           // Discard the tracing span, we don't need it if RUM is tracking the resource
+          tracingSpan.cancel();
           tracingSpan = null;
         }
 
@@ -143,7 +145,8 @@ class DatadogTrackingHttpClient implements HttpClient {
     } catch (e) {
       if (rumKey != null) {
         try {
-          rum?.stopResourceLoadingWithErrorInfo(rumKey, e.toString());
+          rum?.stopResourceLoadingWithErrorInfo(
+              rumKey, e.toString(), e.runtimeType.toString());
         } catch (innerE) {
           datadogSdk.internalLogger.sendToDatadog(
               '$DatadogTrackingHttpClient encountered an error while attempting '
@@ -318,7 +321,8 @@ class _DatadogTrackingHttpRequest implements HttpClientRequest {
   void _onStreamError(Object e, StackTrace? st) {
     try {
       if (rumKey != null) {
-        datadogSdk.rum?.stopResourceLoadingWithErrorInfo(rumKey!, e.toString());
+        datadogSdk.rum?.stopResourceLoadingWithErrorInfo(
+            rumKey!, e.toString(), e.runtimeType.toString());
       }
       tracingContext?.setErrorInfo(
         e.runtimeType.toString(),
@@ -470,8 +474,8 @@ class _DatadogTrackingHttpResponse extends Stream<List<int>>
 
       if (rumKey != null) {
         if (lastError != null) {
-          datadogSdk.rum
-              ?.stopResourceLoadingWithErrorInfo(rumKey!, lastError.toString());
+          datadogSdk.rum?.stopResourceLoadingWithErrorInfo(
+              rumKey!, lastError.toString(), lastError.runtimeType.toString());
         } else {
           var resourceType = resourceTypeFromContentType(headers.contentType);
           var size = innerResponse.contentLength > 0