+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## Unreleased
 
+## 1.0.0-alpha.2
+
+* Cancel spans on DatadogTrackingHttpClient when RUM is enabled (prevent spans
+  from leaking native resources)
+* Remove native view tracking (Activities and Fragments) from Android by default
+* Add support for creating multiple named loggers: `DatadogSdk.createLogger` and
+  `LoggingConfiguration.loggerName`
+* Add support for configuring whether loggers send data to Datadog:
+  `LoggingConfiguration.sendLogsToDatadog`
+* 💥 Breaking! - Removed `DdSdkConfiguration.trackHttpClient`. This has been
+  replaced with a standalone `datadog_tracking_http_client` package.
+* 💥 Breaking! - `DdSdkConfiguration.site` is now a required parameter and no
+  longer defaults to `DatadogSite.us1`
+
 ## 1.0.0-alpha.1
 
 * Support for Logging, Tracing (including Datadog Distributed Tracing) and RUM